Cleaned up search, put filters into separate files, connected categories and status
